解决拿到的payloadSlice错误的问题
**复现步骤**

trident-testsrv+trident

**原因和解决方案**

1. uint转int的方式错误
2. payloadSlice是可选参数

**影响范围**

影响NPB的基本功能

**验证方案**

通过trident验证成功

**涉及分支**

* feature-npb

**检查项**

- [ ] 需要更新dep依赖
- [ ] 是共性问题(代码中存在类似问题)
- [x] 编译通过
- [x] 单元测试通过

/assign @xiangyang

+++ datatype/policy.go
@@ -23,7 +23,7 @@ func (a NpbAction) TunnelId() uint32 {
 }
 
 func (a NpbAction) PayloadSlice() int {
-	return int(uint16(a))
+	return int(int16(a))
 }
 
 func (a NpbAction) String() string {
@@ -33,8 +33,8 @@ func (a NpbAction) String() string {
 	return fmt.Sprintf("{%d@%s}", a.TunnelId(), IpFromUint32(a.TunnelIp()))
 }
 
-func ToNpbAction(ip uint32, id uint32, slice uint32) NpbAction {
-	return NpbAction(ip)<<32 | NpbAction(id)<<16 | NpbAction(slice)
+func ToNpbAction(ip uint32, id uint32, slice int) NpbAction {
+	return NpbAction(ip)<<32 | NpbAction(uint16(id))<<16 | NpbAction(uint16(slice))
 }
 
 const (
